+++ tests/test_models.py
@@ -194,3 +194,34 @@ class MessageTestCase(AppPushTestCaseMixin, TestCase):
 
         self.assertEqual(message_logs_count, 0)
         self.assertEqual(messages_count, 2)
+
+    def test_message_send_to_group_that_has_app_push_and_email_but_skips_app_push_on_one_key(self):
+        # this message key in the group has app push and email
+        Message.objects.create(user=self.user, key='group_with_skip_push', fail_silently=False)
+
+        messages_count = Message.objects.count()
+        self.assertEqual(messages_count, 1)
+
+        process_new_messages()
+        process_new_message_logs()
+
+        message_logs_count = MessageLog.objects.count()
+        messages_count = Message.objects.count()
+
+        self.assertEqual(message_logs_count, 2)
+        self.assertEqual(messages_count, 1)
+
+        # this message key in the group has no app push
+        Message.objects.create(user=self.user, key='group_with_skip_push_2', fail_silently=False)
+
+        messages_count = Message.objects.count()
+        self.assertEqual(messages_count, 2)
+
+        process_new_messages()
+        process_new_message_logs()
+
+        message_logs_count = MessageLog.objects.count()
+        messages_count = Message.objects.count()
+
+        self.assertEqual(message_logs_count, 3)
+        self.assertEqual(messages_count, 2)